Front End Files
We’ve used Html and Php for the front end and also have uploaded the Sql file for the Database. It consists of a web-based dashboard supported by Firebase Realtime Database showing realtime alerts detected from the system. The dashboard consists of two parts which is RDD alerts and RDD History where you can see the Realtime alerts and History of the Alerts.